refactor: separate CLI parsing from composition root
Split clap command definitions into src/cli.rs and move dependency wiring to src/app.rs so main.rs only parses args and returns ExitCode. Align Config/Daemon commands to tuple-style args and delegate config command handling to configuration CLI context.

Made-with: Cursor
