+++ src/contexts/configuration/interface/cli.rs
@@ -1,1 +1,3 @@
 pub mod config;
+
+pub use config::{ConfigArgs, run};
